% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{auc}
\alias{auc}
\title{AUC calculation}
\usage{
auc(x, y, thresh = NULL, dens = 100, sort.x = TRUE)
}
\arguments{
\item{x}{Vector of x positions.}

\item{y}{Vector of y positions.}

\item{thresh}{Threshold below which area is not calculated.}

\item{dens}{By default the data density is artificially increased by adding 100 data points between given adjacent data points.}

\item{By}{default the vectors in x and y are ordered along increasing x because integration makes no sense with unordered data.}
}
\value{
The AUC.
}
\description{
Taken from "flux" package (not able to install package after update of R). Compute the Area Under the Curve (AUC) between two vectors of points.
}
