% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserv_func.R
\name{get_conservedRegion}
\alias{get_conservedRegion}
\title{Retrieve consensus regions for a list of TADs.}
\usage{
get_conservedRegion(signif_dt, all_tad_pos_dt, all_g2t_dt,
  minOverlapBpRatio = 0.8, minIntersectGenes = 3,
  gene_matching_fuse_threshold = 0.8, nCpu = 2, logFile = NULL,
  verbose = TRUE)
}
\arguments{
\item{signif_dt}{Dataframe with the list of TAds that have to be matched (columns: dataset/region).}

\item{all_tad_pos_dt}{Coordinates of the TADs hold in signif_dt (colums:  dataset/chromo/region/start/end).}

\item{all_g2t_dt}{Gene-to-TAD assignment for the TADs hold in signif_dt (columns: dataset/entrezID/chromo/start/end/region/symbol).}

\item{minOverlapBpRatio}{Retain only matches with >= minOverlapBpRatio bp overlap (filter 1).}

\item{minIntersectGenes}{Retain only conserved regions with >= minIntersectGenes genes at the intersect (filter 2).}

\item{gene_matching_fuse_threshold}{Merge conserved regions that have >= gene_matching_fuse_threshold % gene overlap (after filter 2).}

\item{nCpu}{Number of CPU available.}

\item{logFile}{If provided, write logs in this file.}

\item{verbose}{Verbose or not function execution.}
}
\value{
A list of four elements: the all-vs-all matching table ("matching_table"), the conserved regions with the corresponding TADs ("conserved_signif_tads"), the conserved regions with the corresponding intersect genes ("conserved_signif_intersect_genes"), the list of parameters used for the filters  ("parameters").
}
\description{
Retrieve region corresponding to overlapping TADs.
}
