% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcByRegion_func.R
\name{get_FCdownByRegion_v2}
\alias{get_FCdownByRegion_v2}
\title{RatioFC by TADs}
\usage{
get_FCdownByRegion_v2(g2TADdt, DEdt)
}
\arguments{
\item{g2TADdt}{Gene-to-TAD dataframe.}

\item{DEdt}{Differential expression table.}
}
\value{
The ratioFC score.
}
\description{
Function that calculates the ratioFC score for each region.
}
