% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{get_meanCorr}
\alias{get_meanCorr}
\title{Mean intra-TAD correlation}
\usage{
get_meanCorr(gene2tad_dt, exprd_dt, corrMeth = "pearson",
  withDiag = FALSE, nCpu = 2)
}
\arguments{
\item{gene2tad_dt}{Dataframe with gene-to-TAD assignment (required columns: "entrezID" for gene IDs, "region" for TAD IDs).}

\item{corrMeth}{Correlation method (one of "pearson", "kendall"or "spearman"; default:"pearson").}

\item{withDiag}{If correlation with itself should be included (default: FALSE).}

\item{expr_dt}{Dataframe with gene expression values for which to compute correlation (row names should be gene IDs; all gene IDs from gene2tad_dt should be available).}

\item{plotCex}{Number available CPU.}
}
\value{
A 2-column dataframe (region/meanCorr colums).
}
\description{
Function the average pairwise correlations within TADs.
}
