% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations_func.R
\name{get_multiShuffledPositions_vFunct}
\alias{get_multiShuffledPositions_vFunct}
\title{Gene-to-TAD permutations}
\usage{
get_multiShuffledPositions_vFunct(g2TADdt, RNAdt, geneIDlist, nClass,
  withExprClass, TADonly, nSimu, nCpu, aggregFun)
}
\arguments{
\item{g2TADdt}{Gene-to-TAD dataframe}

\item{RNAdt}{Gene expression dataframe}

\item{geneIDlist}{List of gene IDs that should be used}

\item{nClass}{The number of classes of expression in which genes are shuffled}

\item{withExprClass}{If shuffling should take place within classes}

\item{TADonly}{If only genes from TADs should be used}

\item{nSimu}{The number of permutations}

\item{nCpu}{The number of CPUs to use}

\item{aggregFun}{With which function gene expression should be aggregated (across samples)}
}
\value{
The permtuation data
}
\description{
Performs the permutation (parallelized).
}
