% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{get_ratioFC}
\alias{get_ratioFC}
\title{RatioFC score}
\usage{
get_ratioFC(fc_vect)
}
\arguments{
\item{fc_vect}{Vector of fold-changes.}
}
\value{
The ratioFC score.
}
\description{
Function that calculates ratio of negative fold-change (FC) for a vector of FC values.
}
