% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcByRegion_func.R
\name{get_statFromShuffle_para}
\alias{get_statFromShuffle_para}
\title{RatioFC by TADs}
\usage{
get_statFromShuffle_para(DEdt, shuffData, stat, geneIDlist = NULL,
  ncpu = 2, TADonly = F)
}
\arguments{
\item{DEdt}{Differential expression table.}

\item{shuffData}{The permutation data.}

\item{stat}{The score to compute.}

\item{ncpu}{The number of CPUs to use.}

\item{TADonly}{If only genes from TADs should be used.}
}
\value{
The score for each permtuation.
}
\description{
Function that calculates the ratioFC score for each region.
}
