% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_func.R
\name{head_sq}
\alias{head_sq}
\title{Return the first lines/columns of a dataframe}
\usage{
head_sq(dt, ntop = 5)
}
\arguments{
\item{dt}{A two-dimensional dataframe or matrix.}

\item{ntop}{The maximal number of lines/columns to display.}
}
\value{
The first rows/columns of the dataframe.
}
\description{
Returns the first lines and columns of a dataframe (as many columns as lines).
}
