% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_func.R
\name{madNorm}
\alias{madNorm}
\title{Median absolute deviation (MAD) of a dataframe}
\usage{
madNorm(x)
}
\arguments{
\item{x}{Dataframe that will undergo MAD-normalization.}
}
\value{
The normalized dataframe.
}
\description{
Performs MAD-normalization of a dataframe (use for microarray data).
}
