% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_func.R
\name{outTxt}
\alias{outTxt}
\title{Print text and log}
\usage{
outTxt(txt, verbose, logFile)
}
\arguments{
\item{txt}{The text to be printed.}

\item{verbose}{If the text should be "cat"."}

\item{logFile}{File where to append the text (if not NULL)}
}
\description{
Function that "cat" or not a message and print it or not to log file.
}
