% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{plot_conservedRegions}
\alias{plot_conservedRegions}
\title{Plot TADs of a conserved region}
\usage{
plot_conservedRegions(genes_dt, tads_dt, TADlinecol1 = "darkblue",
  TADlinecol2 = "darkturquoise", consAll_col = "darkgreen",
  consAbove_col = "orange", consBelow_col = "black",
  tad_gene_space = 0.5, gene_line_offset = 0.1, TADlt = 1,
  TADlw = 2, geneLt = 1, geneLw = 1, geneDelLt = 2,
  geneDelLw = 0.5, dsCat_cols = NULL, colConsThresh = NULL,
  subTit = NULL, myTit = NULL)
}
\arguments{
\item{genes_dt}{The text to be printed. Should contain at least the following columns: symbol/chromo/start/end/<count>}

\item{tads_dt}{Dataframe with information regarding conserved TADs to plot. Should contain at least the following columns: dataset/dsCat/cond1/cond2/chromo/start/end/<upCond>. Datasets will be grouped according to dsCat.}

\item{dsCat_cols}{Should be a named vector of colors. The names should match the levels of tads_dt$dsCat.}

\item{...}{Other parameters for fine-tuning the plot (might be to look a bit in the code to adapt to your data).}
}
\value{
The plot (ggplot).
}
\description{
Function to plot the TADs and genes of a conserved region.
}
