% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_func.R
\name{printAndLog}
\alias{printAndLog}
\title{Print text and log}
\usage{
printAndLog(mytext, mylogf)
}
\arguments{
\item{mytext}{The text to be printed.}

\item{mylogf}{Location where to print the text (e.g. log file or NULL)}
}
\value{
The ratioFC score
}
\description{
Function that "cat" a message and print it to log file.
}
