% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_func.R
\name{quantNorm}
\alias{quantNorm}
\title{Quantile normalization of a vector}
\usage{
quantNorm(x)
}
\arguments{
\item{x}{Vector that will undergo quantil-normalization.}
}
\value{
The normalized vector.
}
\description{
Performs quantile-normalization of a vector of numeric values.
}
